I had modified from decimal to float

+++ Archaeological site/ArchaeologicalTests.cs
@@ -9,12 +9,12 @@ namespace Archaeological_site
         [TestMethod]
         public void FirstTry()
         {
-            decimal buildingArea = buildingpoints(1.000001m, 2.000002m, 3.000003m, 4.000004m, 6.000006m, 1.000001m);
-            Assert.AreEqual(6.000012m, buildingArea);
+            float buildingArea = buildingpoints(1.000001f, 2.000002f, 3.000003f, 4.000004f, 6.000006f, 1.000001f);
+            Assert.AreEqual(6.000012f, buildingArea);
         }
-        decimal buildingpoints(decimal Xa, decimal Ya, decimal Xb, decimal Yb, decimal Xc, decimal Yc)
+        float buildingpoints(float Xa, float Ya, float Xb, float Yb, float Xc, float Yc)
         {
-            decimal area = Math.Abs((Xa * (Yb - Yc) + Xb * (Yc - Ya) + Xc * (Ya - Yb)) / 2);
+            float area = Math.Abs((Xa * (Yb - Yc) + Xb * (Yc - Ya) + Xc * (Ya - Yb)) / 2);
             return area;
         }
